Fixed issues with caching (#4)

+++ .github/workflows/github_job_scheduler.py
@@ -3,9 +3,57 @@
 import json
 import math
 import argparse
+from enum import IntEnum
 from itertools import chain
 from pathlib import PosixPath
 from typing import List, Optional
+
+
+class State(IntEnum):
+    NORMAL = 0
+    BLOCK_SINGLE_QUOTE = 1
+    BLOCK_DOUBLE_QUOTE = 2
+
+
+def _remove_executable_from_doc(input_file_path: PosixPath, output_file_path: PosixPath) -> None:
+    lines = []
+    current_state = State.NORMAL
+    with input_file_path.open() as fh:
+        for line in fh:
+            if current_state == State.NORMAL:
+                if line.startswith("#"):
+                    lines.append(line)
+                    current_state = State.NORMAL
+                elif line.startswith(('"""', 'r"""')) and line.endswith(('"""', '"""\n')) and len(line) > 6:
+                    lines.append(line)
+                    current_state = State.NORMAL
+                elif line.startswith(('"""', '"""\n', 'r"""', 'r"""\n')):
+                    lines.append(line)
+                    current_state = State.BLOCK_DOUBLE_QUOTE
+                elif line.startswith(("'''", "r'''")) and line.endswith(("'''", "'''\n")) and len(line) > 6:
+                    lines.append(line)
+                    current_state = State.NORMAL
+                elif line.startswith(("'''", "'''\n", "r'''", "r'''\n")):
+                    lines.append(line)
+                    current_state = State.BLOCK_SINGLE_QUOTE
+                else:
+                    lines.append("\n")
+                    current_state = State.NORMAL
+            elif current_state == State.BLOCK_DOUBLE_QUOTE:
+                lines.append(line)
+                current_state = State.NORMAL if line.startswith(('"""', '"""\n')) or line.endswith(('"""', '"""\n')) \
+                    else State.BLOCK_DOUBLE_QUOTE
+            elif current_state == State.BLOCK_SINGLE_QUOTE:
+                lines.append(line)
+                current_state = State.NORMAL if line.startswith(("'''", "'''\n")) or line.endswith(("'''", "'''\n")) \
+                    else State.BLOCK_SINGLE_QUOTE
+    lines.append("\n# %%%%%%RUNNABLE_CODE_REMOVED%%%%%%")
+
+    with output_file_path.open("w", encoding="utf-8") as fh:
+        for line in lines:  # More space efficient in this case than using str.join
+            fh.write(line)
+
+    return None
 
 
 def build_matrix(num_workers: int,
@@ -13,7 +61,7 @@ def build_matrix(num_workers: int,
                  parser_namespace: argparse.Namespace) -> List[str]:
     glob_pattern = parser_namespace.glob_pattern
     file_count = len(list(build_directory.glob(glob_pattern)))
-    files_per_worker = math.ceil(file_count / num_workers)
+    files_per_worker = math.ceil(file_count / (num_workers + 1))
 
     return list(range(1, file_count, files_per_worker))
 
@@ -29,7 +77,7 @@ def execute_matrix(num_workers: int,
     assert offset and offset > 0, f"Invalid value for offset. Expected int greater than 0; Got: {offset}"
 
     files = sorted(build_directory.glob(glob_pattern))
-    files_per_worker = math.ceil(len(files) / num_workers)
+    files_per_worker = math.ceil(len(files) / (num_workers + 1))
 
     files_to_retain = files[offset - 1:offset + files_per_worker - 1]
     if dry_run:
@@ -39,10 +87,41 @@ def execute_matrix(num_workers: int,
     files_to_delete_2 = files[offset + files_per_worker - 1:]
 
     for file in chain(files_to_delete_1, files_to_delete_2):
-        file.unlink()
+        _remove_executable_from_doc(file, file)
         if verbose:
-            print("Deleted", file.name)
+            print("Removing executable code from", file.name)
+    return None
 
+
+def clean_html(num_workers: int,
+               root_directory: PosixPath,
+               build_directory: PosixPath,
+               parser_namespace: argparse.Namespace) -> Optional[List[str]]:
+    current_dry_run = parser_namespace.dry_run
+    verbose = parser_namespace.verbose
+    parser_namespace.dry_run = True
+    files_to_retain = execute_matrix(num_workers, build_directory, parser_namespace)
+    files_to_retain = list(map(lambda f: "".join(f.split(".")[:-1]), files_to_retain))
+
+    html_files = (root_directory / "_build" / "html" / "demos").glob("*.html")
+
+    dry_run = []
+    for file in html_files:
+        file_stem = file.stem
+
+        if file_stem == "index":
+            continue
+
+        if file_stem not in files_to_retain:
+            if current_dry_run:
+                dry_run.append(file.name)
+            else:
+                file.unlink()
+            if verbose:
+                print("Deleted", file.name)
+
+    if current_dry_run:
+        return dry_run
     return None
 
 
@@ -52,9 +131,10 @@ if __name__ == "__main__":
         description="This Python script aids in splitting the build process of the QML docs across multiple nodes"
     )
     parser.add_argument("action",
-                        help="Indicate whether the job schedule has be to built "
-                             "or a built job schedule has to be executed",
-                        choices=["build-matrix", "execute-matrix"],
+                        help="build-matrix -> Build strategy.matrix that will be used to schedule jobs dynamically. "
+                             "execute-matrix -> Remove executable code from non-relevant files in current matrix offset. "
+                             "clean-html -> Delete html files built that are not relevant for current matrix offset.",
+                        choices=["build-matrix", "execute-matrix", "clean-html"],
                         type=str)
     parser.add_argument("directory",
                         help="The path to the qml directory",
@@ -74,7 +154,7 @@ if __name__ == "__main__":
                         default="*.py")
     parser_build = parser.add_argument_group("Build Matrix")
 
-    parser_execute = parser.add_argument_group("Execute Matrix")
+    parser_execute = parser.add_argument_group("Execute Matrix / Clean")
     parser_execute.add_argument("--offset",
                                 help="The current matrix output to retain files from",
                                 type=int)
@@ -95,9 +175,12 @@ if __name__ == "__main__":
     worker_count = parser_results.num_workers
     assert worker_count > 0, "Total number of workers has to be greater than 1"
 
-    output = {
-        "build-matrix": build_matrix,
-        "execute-matrix": execute_matrix
-    }[parser_results.action](worker_count, directory_examples, parser_results)
+    if parser_results.action == "clean-html":
+        output = clean_html(worker_count, directory_qml, directory_examples, parser_results)
+    else:
+        output = {
+            "build-matrix": build_matrix,
+            "execute-matrix": execute_matrix
+        }[parser_results.action](worker_count, directory_examples, parser_results)
 
     print(json.dumps(output) if output else "")
